Add router with splash page and snarky contact form
